Fix my_any
